crm_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
simple-salesforce
